backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
httpx==0.24.1
pydantic==1.10.12
SQLAlchemy==1.4.49
asyncpg==0.27.0
alembic==1.11.1
pytest==7.4.0
pytest-asyncio==0.22.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
aiosqlite==0.18.0


